+++ src/verbs/mod.rs
@@ -122,6 +122,7 @@ pub mod timer;
 pub mod title;
 pub mod trail;
 pub mod transcode;
+pub mod upscale;
 pub mod v360;
 pub mod vdenoise;
 pub mod vhs;